requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2.16
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
pipreqs==0.4.11
pytz==2022.7
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
yarg==0.1.9
